refactor(site): race-weekend schedule fetched server-side
Phase D of the pure-SSR migration.

- move the NextRaceBySeason gql doc + result type into a client-free
  raceWeekend/queries.ts; cached-data getSeasonRaceSchedule(season) fetches it
- selectNextRace/mapToNextRace extracted as pure (todayISO passed in)
- RaceWeekend takes the server-fetched schedule as a prop and computes the
  upcoming race client-side (wall-clock stays client); useNextRaceData browser
  query removed
- season pages (home + /[season]) and the circuit page thread the schedule
  through to RaceWeekend / NextRaceCountdown (no browser query, no cast)

Build compiles; tsc + 39 tests green.
