+++ packages/site/src/app/[season]/page.tsx
@@ -5,6 +5,7 @@ import {
 	getCurrentSeason,
 	getDriverStandings,
 	getSeason,
+	getSeasonRaceSchedule,
 	getSeasonSchedule,
 	getSeasonStats
 } from '../lib/cached-data';
@@ -25,18 +26,26 @@ export async function generateMetadata({ params }: { params: Params }): Promise<
 export default async function SeasonPage({ params }: { params: Params }) {
 	const { season: seasonParam } = await params;
 	const year = Number(seasonParam);
-	const [season, scheduleData, driverStandingsData, constructorStandingsData, statsBundle] =
-		await Promise.all([
-			getSeason(year),
-			getSeasonSchedule(year),
-			getDriverStandings(year),
-			getConstructorStandings(year),
-			getSeasonStats(year)
-		]);
+	const [
+		season,
+		races,
+		scheduleData,
+		driverStandingsData,
+		constructorStandingsData,
+		statsBundle
+	] = await Promise.all([
+		getSeason(year),
+		getSeasonRaceSchedule(year),
+		getSeasonSchedule(year),
+		getDriverStandings(year),
+		getConstructorStandings(year),
+		getSeasonStats(year)
+	]);
 
 	return (
 		<SeasonContent
 			season={season}
+			races={races}
 			scheduleData={scheduleData}
 			driverStandingsData={driverStandingsData}
 			constructorStandingsData={constructorStandingsData}
